refactor: isolate schema ownership support code
